+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Changes
 
 * [#7181](https://github.com/rubocop-hq/rubocop/pull/7181): Sort analyzed file alphabetically. ([@pocke][])
+* [#7188](https://github.com/rubocop-hq/rubocop/pull/7188): Include inspected file location in auto-correction error. ([@pocke][])
 
 ## 0.72.0 (2019-06-25)